chore(meshupgrade): extrapolate stepper logic
