+++ plugins/lime-plugin-mesh-wide-upgrade/src/hooks/meshWideUpgradeProvider.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from "preact/hooks";
 import { useCallback, useContext } from "react";
 
 import { useNewVersion } from "plugins/lime-plugin-firmware/src/firmwareQueries";
+import { getStepperStatus } from "plugins/lime-plugin-mesh-wide-upgrade/src/hooks/use-stepper";
 import {
     meshUpgradeNodeStatusKey,
     useBecomeMainNode,
@@ -18,7 +19,6 @@ import {
     StepperState,
 } from "plugins/lime-plugin-mesh-wide-upgrade/src/meshUpgradeTypes";
 import { getMeshWideError } from "plugins/lime-plugin-mesh-wide-upgrade/src/utils/processError";
-import { getStepperStatus } from "plugins/lime-plugin-mesh-wide-upgrade/src/utils/stepper";
 
 import { useBoardData, useSession } from "utils/queries";
 import queryCache from "utils/queryCache";
